+++ src/widget/animationcontrolwidget.cpp
@@ -1,3 +1,4 @@
+#include <cmath>
 #include <QTime>
 #include <QWidget>
 #include "animationcontrolwidget.h"
@@ -157,13 +158,26 @@ namespace PixelMaestroStudio {
 		ui->cycleTimeEdit->setTime(QTime::fromMSecsSinceStartOfDay(value));
 		ui->cycleTimeEdit->blockSignals(false);
 
+		ui->cycleSpinBox->blockSignals(true);
+		ui->cycleSpinBox->setValue(Timer::millis_to_upm(value));
+		ui->cycleSpinBox->blockSignals(false);
+
 		set_animation_timer();
 	}
 
+	void PixelMaestroStudio::AnimationControlWidget::on_cycleSpinBox_editingFinished() {
+		ui->cycleIntervalSlider->setValue(Timer::upm_to_millis(ui->cycleSpinBox->value()));
+	}
+
 	void PixelMaestroStudio::AnimationControlWidget::on_cycleTimeEdit_editingFinished() {
+		int millis = ui->cycleTimeEdit->time().msecsSinceStartOfDay();
 		ui->cycleIntervalSlider->blockSignals(true);
-		ui->cycleIntervalSlider->setValue(ui->cycleTimeEdit->time().msecsSinceStartOfDay());
+		ui->cycleIntervalSlider->setValue(millis);
 		ui->cycleIntervalSlider->blockSignals(false);
+
+		ui->cycleSpinBox->blockSignals(true);
+		ui->cycleSpinBox->setValue(Timer::millis_to_upm(millis));
+		ui->cycleSpinBox->blockSignals(false);
 
 		set_animation_timer();
 	}
@@ -177,13 +191,27 @@ namespace PixelMaestroStudio {
 		ui->delayTimeEdit->setTime(QTime::fromMSecsSinceStartOfDay(value));
 		ui->delayTimeEdit->blockSignals(false);
 
+		ui->delaySpinBox->blockSignals(true);
+		ui->delaySpinBox->setValue(Timer::millis_to_upm(value));
+		ui->delaySpinBox->blockSignals(false);
+
 		set_animation_timer();
 	}
 
+	void AnimationControlWidget::on_delaySpinBox_editingFinished() {
+		ui->delayIntervalSlider->setValue(Timer::upm_to_millis(ui->delaySpinBox->value()));
+	}
+
 	void AnimationControlWidget::on_delayTimeEdit_editingFinished() {
+		int millis = ui->delayTimeEdit->time().msecsSinceStartOfDay();
+
 		ui->delayIntervalSlider->blockSignals(true);
-		ui->delayIntervalSlider->setValue(ui->delayTimeEdit->time().msecsSinceStartOfDay());
+		ui->delayIntervalSlider->setValue(millis);
 		ui->delayIntervalSlider->blockSignals(false);
+
+		ui->delaySpinBox->blockSignals(true);
+		ui->delaySpinBox->setValue(Timer::millis_to_upm(millis));
+		ui->delaySpinBox->blockSignals(false);
 
 		set_animation_timer();
 	}
@@ -295,8 +323,10 @@ namespace PixelMaestroStudio {
 			ui->reverseCheckBox->blockSignals(true);
 			ui->fadeCheckBox->blockSignals(true);
 			ui->cycleIntervalSlider->blockSignals(true);
+			ui->cycleSpinBox->blockSignals(true);
 			ui->cycleTimeEdit->blockSignals(true);
 			ui->delayIntervalSlider->blockSignals(true);
+			ui->delaySpinBox->blockSignals(true);
 			ui->delayTimeEdit->blockSignals(true);
 			ui->centerXSpinBox->blockSignals(true);
 			ui->centerYSpinBox->blockSignals(true);
@@ -304,8 +334,10 @@ namespace PixelMaestroStudio {
 			ui->reverseCheckBox->setChecked(animation->get_reverse());
 			ui->fadeCheckBox->setChecked(animation->get_fade());
 			ui->cycleIntervalSlider->setValue(animation->get_timer()->get_interval());
+			ui->cycleSpinBox->setValue(Timer::millis_to_upm(animation->get_timer()->get_interval()));
 			ui->cycleTimeEdit->setTime(QTime::fromMSecsSinceStartOfDay(animation->get_timer()->get_interval()));
 			ui->delayIntervalSlider->setValue(animation->get_timer()->get_delay());
+			ui->delaySpinBox->setValue(Timer::millis_to_upm(animation->get_timer()->get_delay()));
 			ui->delayTimeEdit->setTime(QTime::fromMSecsSinceStartOfDay(animation->get_timer()->get_delay()));
 			ui->centerXSpinBox->setValue(animation->get_center().x);
 			ui->centerYSpinBox->setValue(animation->get_center().y);
@@ -313,8 +345,10 @@ namespace PixelMaestroStudio {
 			ui->reverseCheckBox->blockSignals(false);
 			ui->fadeCheckBox->blockSignals(false);
 			ui->cycleIntervalSlider->blockSignals(false);
+			ui->cycleSpinBox->blockSignals(false);
 			ui->cycleTimeEdit->blockSignals(false);
 			ui->delayIntervalSlider->blockSignals(false);
+			ui->delaySpinBox->blockSignals(false);
 			ui->delayTimeEdit->blockSignals(false);
 			ui->centerXSpinBox->blockSignals(false);
 			ui->centerYSpinBox->blockSignals(false);
@@ -433,6 +467,10 @@ namespace PixelMaestroStudio {
 		ui->centerXSpinBox->setEnabled(enabled);
 		ui->centerYSpinBox->setEnabled(enabled);
 		ui->currentCycleSpinBox->setEnabled(ui->playbackStartStopToolButton->isChecked());
+	}
+
+	void AnimationControlWidget::enable_cpm_timing(bool enable_cpm) {
+		this->enable_cpm_ = enable_cpm;
 	}
 
 	AnimationControlWidget::~AnimationControlWidget() {